+++ frontend/src/main.tsx
@@ -3,10 +3,13 @@ import ReactDOM from 'react-dom/client'
 import App from './App.tsx'
 import './index.css'
 import { Toaster } from 'sonner'
+import { ImagePreviewProvider } from '@/components/image-preview/ImagePreviewContext'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <App />
+    <ImagePreviewProvider>
+      <App />
+    </ImagePreviewProvider>
     <Toaster 
       position="top-right"
       richColors